+++ react-components/src/App.tsx
@@ -5,7 +5,7 @@ import { Routes, Route } from 'react-router-dom';
 import About from './Pages/About/About';
 import Main from './Pages/Main/Main';
 import Page404 from './Pages/Page404/Page404';
-import { MyFormContainer } from './Pages/FormPage/FormPage';
+import { FormPage } from './Pages/FormPage/FormPage';
 
 function App() {
   return (
@@ -13,7 +13,7 @@ function App() {
       <Header />
       <Routes>
         <Route path="/" element={<Main />} />
-        <Route path="/form" element={<MyFormContainer />} />
+        <Route path="/form" element={<FormPage />} />
         <Route path="/about" element={<About />} />
         <Route path="*" element={<Page404 />} />
       </Routes>
